various: use more optimized brute force algorithm
This algorithm guarantees that if a result exists, it will be found, and in
the least possible iterations, with no duplicate results.

It's goal is to find an encrypted series of bytes at a specified location in the
Pokémon's data. This is done by seeing this byteseries as an encrypted state,
and for each checksum attempting to decrypt it, initially disregarding the rest
of the data of the Pokémon.

This will very likely result in a decrypted state that is not valid, so
therefore a series of checks is ran on the decrypted state to determine whether
this is a valid input.

If it is, then the next step is to take the original provided Pokémon data
and set that input data, and calculating the current checksum. We also provide
a series of easy to modify parameters, and brute force the difference in checksum
this will generate. At that point, we simply look through the list as a lookup table
to find if we can in fact manipulate the checksum to be the desired one so that the
encrypted state matches our initial search goal.

Signed-off-by: Jorik Devreese <jorikdevreese@hotmail.be>